+++ bin/stable/prepare_project_edition.sh
@@ -101,7 +101,11 @@ fi
 sudo sed -i "s/\['test' => true\]/\['test' => true, 'behat' => true\]/g" config/bundles.php
 
 # Create a default Behat configuration file
-cp "behat_ibexa_${PROJECT_EDITION}.yaml" behat.yaml
+if [ -f "behat_ibexa_${PROJECT_EDITION}.php" ]; then
+    cp "behat_ibexa_${PROJECT_EDITION}.php" behat.php
+else
+    cp "behat_ibexa_${PROJECT_EDITION}.yaml" behat.yaml
+fi
 
 # Depenencies are installed and container can be removed
 docker container stop install_dependencies
